refactoring remove nosql

+++ elecdb-server/src/main/java/net/homenet/dao/repository/ICategoryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author denbilyk
  *         Created: 5/31/16
  */
-public interface ICategoryRepository extends JpaRepository<CategoryRecord, Long> {
+public interface ICategoryRepository extends JpaRepository<CategoryRecord, Integer> {
 
     CategoryRecord findCategoryRecordByCategory(String category);
